Renamed groupId from 'red.zyc' to 'io.allurx' and updated package structure accordingly.
