+++ src/main/java/module-info.java
@@ -1,9 +1,27 @@
-/**
- * @author zyc
+/*
+ * Copyright 2024 allurx
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
-module desensitization {
-    exports red.zyc.desensitization;
-    exports red.zyc.desensitization.annotation;
-    exports red.zyc.desensitization.handler;
-    requires annotation.parser;
+/**
+ * blur module
+ *
+ * @author allurx
+ */
+module io.allurx.blur {
+    exports io.allurx.blur;
+    exports io.allurx.blur.annotation;
+    exports io.allurx.blur.handler;
+    requires io.allurx.kit.base;
+    requires io.allurx.annotation.parser;
 }
