--- src/main/java/red/zyc/desensitization/annotation/Password.java
@@ -1,66 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package red.zyc.desensitization.annotation;
-
-import red.zyc.desensitization.handler.PasswordHandler;
-import red.zyc.parser.handler.Parse;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * 密码敏感注解。默认的脱敏规则：擦除目标对象中所有的字符。
- *
- * @author zyc
- */
-@Target({ElementType.FIELD, ElementType.TYPE_USE, ElementType.PARAMETER, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Parse(handler = PasswordHandler.class, annotation = Password.class)
-public @interface Password {
-
-    /**
-     * @return 敏感信息在原字符序列中的起始偏移
-     */
-    int startOffset() default 0;
-
-    /**
-     * @return 敏感信息在原字符序列中的结束偏移
-     */
-    int endOffset() default 0;
-
-    /**
-     * 只要regexp不为{@code ""}就会忽略{@link #startOffset()}和{@link #endOffset()}
-     *
-     * @return 匹配敏感信息的正则表达式
-     */
-    String regexp() default "";
-
-    /**
-     * @return 敏感信息替换后的占位符
-     */
-    char placeholder() default '*';
-
-    /**
-     * @return 是否需要对目标对象进行脱敏的条件
-     */
-    Class<? extends Condition<?>> condition() default AlwaysTrue.class;
-
-}
