--- src/main/java/red/zyc/desensitization/handler/ChineseNameHandler.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2023 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package red.zyc.desensitization.handler;
-
-import red.zyc.desensitization.annotation.ChineseName;
-
-/**
- * 中文名称处理器
- *
- * @author zyc
- */
-public class ChineseNameHandler extends AbstractCharSequenceHandler<String, ChineseName> {
-
-    @Override
-    public String handle(String target, ChineseName annotation) {
-        return required(target, annotation.condition()) ? String.valueOf(desensitize(target, annotation.regexp(), annotation.startOffset(), annotation.endOffset(), annotation.placeholder())) : target;
-    }
-
-}
